+++ example/src/views/pages/show/index.jsx
@@ -26,8 +26,7 @@ export default React.createClass({
 
   _onRemoveList() {
     let { app, params } = this.props
-    app.push(ListActions.remove, params.id)
-    this.goHome()
+    app.push(ListActions.remove, params.id, this.goHome)
   }
 
 })