+++ src/__tests__/Store-test.js
@@ -2,13 +2,9 @@ import Store from '../Store'
 
 describe('Store', function() {
 
-  it ('defaults to getInitialState within the deserialize method', function() {
-    let store = Object.create(Store)
-
-    store.getInitialState = sinon.mock()
-    store.deserialize()
-
-    store.getInitialState.should.have.been.called
+  it ('adds a toString method', function() {
+    let store = new Store({}, 'sample')
+    store.toString().should.equal('sample')
   })
 
 })